backend/requirements.txt
# Core Dependencies
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-filter==25.1
django-allauth==65.8.1
django-guardian==3.0.0
django-otp==1.6.0
django-storages==1.14.6
django-cors-headers==4.3.1
django-cleanup==8.0.0
django-prometheus==2.3.1
django-health-check==3.17.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
drf-spectacular==0.28.0
django-simple-history==3.4.0

# Database
psycopg2-binary==2.9.10
psycopg==3.2.9
psycopg-binary==3.2.9

# Cache and Task Queue
redis==6.2.0
celery==5.5.2
flower==2.0.1

# Search
elasticsearch==8.18.1
elasticsearch-dsl==8.18.0

# Storage
boto3==1.38.25
botocore==1.38.25

# Security
PyJWT==2.9.0
cryptography==42.0.5

# Documentation and API
drf-yasg==1.21.7
markdown==3.5.2

# PDF Processing
weasyprint==65.1
PyMuPDF==1.26.0
pytesseract==0.3.13

# Utilities
python-decouple==3.8
Pillow==11.2.1
pycountry==24.6.1
humanize==4.12.3

# Development and Testing
black==25.1.0
flake8==7.2.0
pytest==8.3.5
pytest-django==4.11.1
coverage==7.8.2
ipython==9.2.0

# Monitoring and Logging
prometheus-client==0.22.0

# Dependencies of dependencies
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
decorator==5.2.1
executing==2.2.0
inflection==0.5.1
jedi==0.19.2
jsonschema==4.24.0
kombu==5.5.3
packaging==25.0
parso==0.8.4
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
urllib3==2.4.0
vine==5.1.0
